requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic-settings
python-jose[cryptography]
python-multipart
alembic
passlib[argon2]
pytest
httpx
jinja2
